Database connection process moved to seperate file
