+++ server.js
@@ -1,6 +1,5 @@
 const express = require("express");
 require("dotenv").config();
-const mysql = require("mysql");
 
 const PORT = process.env.PORT;
 
@@ -14,20 +13,8 @@ app.use(
   })
 );
 
-const connection = mysql.createConnection({
-  host: process.env.HOST,
-  user: process.env.USER,
-  password: process.env.PASSWORD,
-  database: process.env.DATABASE,
-});
-
-connection.connect((err) => {
-  if (err) {
-    console.error(`Error connecting to databses : ${err}`);
-    return;
-  }
-  console.log("Database successfully connected");
-});
+const connection = require("./Utils/DatabaseConnection").connectDB();
+const SQLExecuter = require("./Utils/SQLExecuter");
 
 app.get("/in", (req, res) => {
   let array = req.query.key.replace(/\[|\]/g, "").split(",");
@@ -37,13 +24,7 @@ app.get("/in", (req, res) => {
 
 app.get("/product/:inqury", (req, res) => {
   if (req.params.inqury == "home") {
-    connection.query("SELECT * FROM Product", (err, result) => {
-      if (err) {
-        console.error(`Error fetching data : ${err}`);
-        return res.status(500).json({ error: "Failed to execute query" });
-      }
-      res.json(result);
-    });
+    SQLExecuter.query(res, connection, "SELECT * FROM Product");
   }
 });
